+++ tests/integration/test_metrics.py
@@ -310,7 +310,7 @@ def test_metrics_max_batch_size(mini_sentry, relay, max_batch_size, expected_eve
     assert mini_sentry.captured_envelopes.empty()
 
 
-@pytest.mark.parametrize("ns", [None, "custom", "spans"])
+@pytest.mark.parametrize("ns", [None, "transactions", "spans"])
 def test_metrics_rate_limits_namespace(mini_sentry, relay, ns):
     relay = relay(mini_sentry, options=TEST_CONFIG)
 
@@ -468,11 +468,10 @@ def test_metrics_with_processing(mini_sentry, relay_with_processing, metrics_con
     metrics_consumer = metrics_consumer()
 
     project_id = 42
-    project_config = mini_sentry.add_full_project_config(project_id)
-    project_config["config"]["features"] = ["organizations:custom-metrics"]
+    mini_sentry.add_full_project_config(project_id)
 
     timestamp = int(datetime.now(tz=timezone.utc).timestamp())
-    metrics_payload = f"spans/foo:42|c\nbar@second:17|c|T{timestamp}"
+    metrics_payload = f"spans/foo:42|c\ntransactions/bar@second:17|c|T{timestamp}"
     relay.send_metrics(project_id, metrics_payload)
 
     metrics = metrics_by_name(metrics_consumer, 2)
@@ -490,12 +489,14 @@ def test_metrics_with_processing(mini_sentry, relay_with_processing, metrics_con
         "received_at": time_after(timestamp),
     }
 
-    assert metrics["headers"]["c:custom/bar@second"] == [("namespace", b"custom")]
-    assert metrics["c:custom/bar@second"] == {
+    assert metrics["headers"]["c:transactions/bar@second"] == [
+        ("namespace", b"transactions")
+    ]
+    assert metrics["c:transactions/bar@second"] == {
         "org_id": 1,
         "project_id": project_id,
         "retention_days": 90,
-        "name": "c:custom/bar@second",
+        "name": "c:transactions/bar@second",
         "tags": {},
         "value": 17.0,
         "type": "c",
@@ -517,11 +518,12 @@ def test_global_metrics_with_processing(
     metrics_consumer = metrics_consumer()
 
     project_id = 42
-    project_config = mini_sentry.add_full_project_config(project_id)
-    project_config["config"]["features"] = ["organizations:custom-metrics"]
+    mini_sentry.add_full_project_config(project_id)
 
     timestamp = int(datetime.now(tz=timezone.utc).timestamp())
-    metrics_payload = f"spans/foo:42|c|T{timestamp}\nbar@second:17|c|T{timestamp}"
+    metrics_payload = (
+        f"spans/foo:42|c|T{timestamp}\ntransactions/bar@second:17|c|T{timestamp}"
+    )
     relay.send_metrics(project_id, metrics_payload)
 
     metrics = metrics_by_name(metrics_consumer, 2)
@@ -539,12 +541,14 @@ def test_global_metrics_with_processing(
         "received_at": time_after(timestamp),
     }
 
-    assert metrics["headers"]["c:custom/bar@second"] == [("namespace", b"custom")]
-    assert metrics["c:custom/bar@second"] == {
+    assert metrics["headers"]["c:transactions/bar@second"] == [
+        ("namespace", b"transactions")
+    ]
+    assert metrics["c:transactions/bar@second"] == {
         "org_id": 1,
         "project_id": project_id,
         "retention_days": 90,
-        "name": "c:custom/bar@second",
+        "name": "c:transactions/bar@second",
         "tags": {},
         "value": 17.0,
         "type": "c",
@@ -1007,24 +1011,6 @@ def test_generic_metric_extraction(mini_sentry, relay):
     } in metrics
 
 
-def test_custom_metrics_disabled(mini_sentry, relay_with_processing, metrics_consumer):
-    relay = relay_with_processing(options=TEST_CONFIG)
-    metrics_consumer = metrics_consumer()
-
-    project_id = 42
-    mini_sentry.add_full_project_config(project_id)
-    # NOTE: "organizations:custom-metrics" missing from features
-
-    timestamp = int(datetime.now(tz=timezone.utc).timestamp())
-    metrics_payload = f"spans/foo:42|c\nbar@second:17|c|T{timestamp}"
-    relay.send_metrics(project_id, metrics_payload)
-
-    metrics = metrics_by_name(metrics_consumer, 1)
-
-    assert "c:spans/foo@none" in metrics
-    assert "c:custom/bar@second" not in metrics
-
-
 @pytest.mark.parametrize("is_processing_relay", (False, True))
 @pytest.mark.parametrize(
     "global_generic_filters",
@@ -1207,19 +1193,16 @@ def test_metrics_received_at(
         )
 
     project_id = 42
-    project_config = mini_sentry.add_basic_project_config(project_id)
-    project_config["config"]["features"] = [
-        "organizations:custom-metrics",
-    ]
+    mini_sentry.add_basic_project_config(project_id)
 
     timestamp = int(datetime.now(tz=timezone.utc).timestamp())
-    relay.send_metrics(project_id, "custom/foo:1337|d")
+    relay.send_metrics(project_id, "spans/foo:1337|d")
 
     metric, _ = metrics_consumer.get_metric()
     assert metric == {
         "org_id": 0,
         "project_id": 42,
-        "name": "d:custom/foo@none",
+        "name": "d:spans/foo@none",
         "type": "d",
         "value": [1337.0],
         "timestamp": time_after(timestamp),